requirements.txt
boto3
dj-database-url
django-durationwidget
django-ipware
django-storages
django-widget-tweaks
gunicorn
psycopg2
social-auth-app-django
twilio
whitenoise
